please rename the app to Swarg Recipes
lets make one more change.

in the home page for user, please remove asking for name and language at the start of tthe recipe and lets load a popup asking for name and language. only when the name is either Shobha, Pavan, Kiran, Swetha, Raushan we proceed with the removal of the popup. else we leave the pop on and not let the user access the website. this is for security. since most of them dont have emails. we will use the name for authentication

after making this change, please also list all the implemented features in this webapp