Could you complete the implementation of the recipe logs.
Every recipe exeuction should be logged along with username, start time and end time, and if it was executed till the last step

also store the usernames in the database.

let the last step in teh recipe be that of taking a photo of the finished product along with asking the user for the weight of the project or the no of preps. in this final data collect post the execution of the recipe, also include a wasted option which means the recipe did not turn out well and the product is wasted